Add files via upload
The project entails developing an application especially for people with Parkinson's Disease (PD), a crippling neurodegenerative disorder that affects millions of people worldwide. Our aim is to create a tool that serves as an early detection system, a gait analyzer, and a lifestyle recommendation system for individuals living with PD.

This presents a newly developed business model that focuses on addressing the widespread and incapacitating neurodegenerative condition known as Parkinson's Disease (PD). Parkinson's disease (PD), a condition that affects a significant number of individuals globally, initially manifests as neurological symptoms such as freezing of gait, impaired balance, and instances of falling. Over time, the disease progresses to encompass additional nonmotor manifestations that affect various systems within the body. The model has been specifically developed to cater to households that include elderly individuals, who are particularly affected by Parkinson's disease (PD), as well as individuals with a familial predisposition to PD, as they face an increased susceptibility to the condition.

The importance of early detection and accurate analysis of PD symptoms cannot be overstated. By leveraging the power of AI and machine learning, we can potentially change the lives of millions of PD patients and their families. This is not just a business idea but also a societal mission: to give PD patients a better, more empowered life.

My work as a team member is to do the prototype selection which is feasible,viable and monetizable and the business modelling